Move `convert_to_array()` to utils
